+++ changelog.md
@@ -1,3 +1,7 @@
+## 0.2
+* Represent games and sessions with classes
+* Store and accumulate played time temporarily
+* Print played time
 ## 0.1
 * Detect games flexibly with process names allowing regex
 * Ignore standard binary extensions in process names